backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.47.0
python-dotenv==1.0.1
mistralai==1.2.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.22
argon2-cffi==23.1.0
slowapi==0.1.9
pdfplumber==0.11.4
